Add membership to the solr index
As discussed in liqd/adhocracy#190, option 1, membership is added as a
separate entity to the solr index.

User activity information is stored with the membership entity instead
of the user entity.

This is merely a stash commit, as this implementation variant isn't
used, but might be useful, if memberships shall be added to the solr
index for another reason at another time.

The reason to not use this variant and to go with multiple activity /
rating fields in the user entity is slightly less redundancy and less
documents in the solr index.