Refactor login dialog
This modifies the login dialog in a way that local login and external
logins (openid, facebook and shibboleth) are displayed in a unified
dialog.

This allows to show the external dialogs upfront, but nevertheless
allows to display a warning which is being noticed.

By the way, it simplifies the login code quite a lot by throwing away
code and awkwardness coming from the openid selector.

UPGRADE NOTE: You can now explicitly select which of the OpenID
providers should be shown by default. Configure adhocracy.login_type if
you want to do so.